+++ sentinel-platform-core/src/main/java/com/sentinel/strategy/RiskScoringStrategy.java
@@ -3,5 +3,5 @@ package com.sentinel.strategy;
 import com.sentinel.dto.DeviceEventRequest;
 
 public interface RiskScoringStrategy {
-    int calculateScore(DeviceEventRequest request);
+    double calculateScore(DeviceEventRequest request);
 }
